refactor(form-elements): implement utility function to all form elements
