+++ libs/core/src/components/pds-checkbox/readme.md
@@ -7,18 +7,18 @@
 
 ## Properties
 
-| Property                   | Attribute        | Description                                                                                                                                                                                                                                               | Type      | Default     |
-| -------------------------- | ---------------- | --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | --------- | ----------- |
-| `checked`                  | `checked`        | It determines whether or not the checkbox is checked.                                                                                                                                                                                                     | `boolean` | `undefined` |
-| `componentId` _(required)_ | `component-id`   | String used for checkbox `id` attribute and label `for` attribute.                                                                                                                                                                                        | `string`  | `undefined` |
-| `disabled`                 | `disabled`       | It determines whether or not the checkbox is disabled.                                                                                                                                                                                                    | `boolean` | `undefined` |
-| `helperMessage`            | `helper-message` | String used for helper message below checkbox.                                                                                                                                                                                                            | `string`  | `undefined` |
-| `indeterminate`            | `indeterminate`  | If `true`, the checkbox will visually appear as indeterminate. Only JavaScript can set the objects `indeterminate` property. See [MDN Web Docs](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/checkbox#indeterminate_state_checkboxes). | `boolean` | `undefined` |
-| `invalid`                  | `invalid`        | It determines whether or not the checkbox is invalid.                                                                                                                                                                                                     | `boolean` | `undefined` |
-| `label`                    | `label`          | String used for label text next to checkbox.                                                                                                                                                                                                              | `string`  | `undefined` |
-| `name`                     | `name`           | String used for checkbox `name` attribute.                                                                                                                                                                                                                | `string`  | `undefined` |
-| `required`                 | `required`       | It determines whether or not the checkbox is required.                                                                                                                                                                                                    | `boolean` | `undefined` |
-| `value`                    | `value`          | The value of the checkbox that is submitted with a form.                                                                                                                                                                                                  | `string`  | `undefined` |
+| Property        | Attribute        | Description                                                                                                                                                                                                                                               | Type      | Default     |
+| --------------- | ---------------- | --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | --------- | ----------- |
+| `checked`       | `checked`        | It determines whether or not the checkbox is checked.                                                                                                                                                                                                     | `boolean` | `undefined` |
+| `componentId`   | `component-id`   | String used for checkbox `id` attribute and label `for` attribute.                                                                                                                                                                                        | `string`  | `undefined` |
+| `disabled`      | `disabled`       | It determines whether or not the checkbox is disabled.                                                                                                                                                                                                    | `boolean` | `undefined` |
+| `helperMessage` | `helper-message` | String used for helper message below checkbox.                                                                                                                                                                                                            | `string`  | `undefined` |
+| `indeterminate` | `indeterminate`  | If `true`, the checkbox will visually appear as indeterminate. Only JavaScript can set the objects `indeterminate` property. See [MDN Web Docs](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/checkbox#indeterminate_state_checkboxes). | `boolean` | `undefined` |
+| `invalid`       | `invalid`        | It determines whether or not the checkbox is invalid.                                                                                                                                                                                                     | `boolean` | `undefined` |
+| `label`         | `label`          | String used for label text next to checkbox.                                                                                                                                                                                                              | `string`  | `undefined` |
+| `name`          | `name`           | String used for checkbox `name` attribute.                                                                                                                                                                                                                | `string`  | `undefined` |
+| `required`      | `required`       | It determines whether or not the checkbox is required.                                                                                                                                                                                                    | `boolean` | `undefined` |
+| `value`         | `value`          | The value of the checkbox that is submitted with a form.                                                                                                                                                                                                  | `string`  | `undefined` |
 
 
 ## Events
