+++ libs/core/src/utils/utils.ts
@@ -24,9 +24,10 @@ export const messageId = (id: string, messageType: string) => {
  * Assign aria-description id to relate messages with form element
  */
 export const assignDescription = (id: string, invalid: boolean, helperMessage: string) => {
+  if (!helperMessage) return
+
   let relatedId = messageId(id, 'helper')
 
-  if (!invalid || !helperMessage) return;
   if (invalid) relatedId = messageId(id, 'error');
 
   return relatedId;
